backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt>=4.0.0,<5.0.0
python-multipart
python-dotenv
httpx
pandas
openpyxl
reportlab
